+++ src/main/scala/org/podval/store/Path.scala
@@ -1,6 +1,15 @@
 package org.podval.store
 
-type Path = Seq[Store]
+import org.podval.metadata.Language
+
+final case class Path(stores: Seq[Store]) derives CanEqual:
+  def isEmpty: Boolean = stores.isEmpty
+  def last: Store = stores.last
+  def lastAs[T]: T = stores.last.asInstanceOf[T]
+  def :+(store: Store): Path = Path(stores :+ store)
+  def structureNames: Seq[String] = stores.map(_.names.doFind(Language.English.toSpec).name)
+  def toUrl: String = if stores.isEmpty then "/" else "/" + structureNames.mkString("/")
+  override def toString: String = toUrl
 
 object Path:
-  def last[T](path: Path): T = path.last.asInstanceOf[T]
+  val empty: Path = Path(Seq.empty)