+++ src/main/scala/org/podval/store/Terminal.scala
@@ -1,3 +1,0 @@
-package org.podval.store
-
-trait Terminal extends Store