+++ src/main/scala/org/podval/xml/CssClass.scala
@@ -1,4 +1,3 @@
 package org.podval.xml
 
-/** A CSS class token (not the HTML `class` attribute; that is [[HtmlClass]]). */
 open class CssClass(val name: String)